+++ prism-core/src/main/resources/sql/mysql/prism_get_or_create_action.sql
@@ -1,5 +1,5 @@
 CREATE PROCEDURE %prefix%get_or_create_action
-    (IN `actionKey` VARCHAR(25), OUT `actionId` TINYINT)
+    (IN `actionKey` VARCHAR(25), OUT `actionId` INT)
 BEGIN
     SELECT action_id INTO `actionId` FROM %prefix%actions WHERE action = `actionKey`;
     IF `actionId` IS NULL THEN
